fix(mp3): keep the VBR tag parser header-only
Reverts the .cpp.hpp split from the previous commit. The repo's linter hard-bans
including a .cpp.hpp from anything but a _build.* file -- "no opt-out" -- and
the split put exactly that include in mp3_vbr_tag.h, breaking cpp_linting.

I went looking for precedent before making the split and found that no FastLED
header includes a .cpp.hpp. I read that as "no strong convention here" when it
was the opposite: there is no precedent because the pattern is forbidden. The
absence was the evidence, and I took it the wrong way round.

The header-only form is what the constraints actually leave. This parser has
three consumers -- the streaming decoder, the unit tests, and the CI
conformance harness -- and the harness builds against `-I src` with no FastLED
library to link, so it cannot reach the definitions through a build file. The
functions are inline with no function-local statics, so the concerns behind the
src/**/*.h rule do not apply. That reasoning is now recorded in the header so
the next reader does not repeat the round trip.

The fl::span<const fl::u8> signature from that commit is kept: it is a real
improvement and unrelated to where the definitions live.

Verified: fl_codec unit tests, conformance 69 passed / 0 failed, cpp_linting
green.
